+++ test/fileurlreader.cpp
@@ -5,9 +5,9 @@
 #include <unistd.h>
 
 #include "3rd-party/catch.hpp"
-#include "test-helpers/chmod.h"
-#include "test-helpers/misc.h"
-#include "test-helpers/tempfile.h"
+#include "test_helpers/chmod.h"
+#include "test_helpers/misc.h"
+#include "test_helpers/tempfile.h"
 #include "utils.h"
 
 using namespace newsboat;
