+++ test/opmlurlreader.cpp
@@ -5,8 +5,8 @@
 #include <unistd.h>
 
 #include "3rd-party/catch.hpp"
-#include "test-helpers/misc.h"
-#include "test-helpers/tempfile.h"
+#include "test_helpers/misc.h"
+#include "test_helpers/tempfile.h"
 #include "utils.h"
 
 using namespace newsboat;
