+++ test/queueloader.cpp
@@ -3,9 +3,9 @@
 #include <fstream>
 
 #include "3rd-party/catch.hpp"
-#include "test-helpers/chmod.h"
-#include "test-helpers/misc.h"
-#include "test-helpers/tempfile.h"
+#include "test_helpers/chmod.h"
+#include "test_helpers/misc.h"
+#include "test_helpers/tempfile.h"
 
 #include "configcontainer.h"
 #include "download.h"
