+++ test/rssitem.cpp
@@ -7,8 +7,8 @@
 #include "configcontainer.h"
 #include "rssfeed.h"
 
-#include "test-helpers/envvar.h"
-#include "test-helpers/stringmaker/optional.h"
+#include "test_helpers/envvar.h"
+#include "test_helpers/stringmaker/optional.h"
 
 using namespace newsboat;
